src/requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.8
# by the following command:
#
#    pip-compile src/requirements.in
#
asgiref==3.5.0
    # via django
async-timeout==4.0.2
    # via redis
attrs==21.4.0
    # via jsonschema
babel==2.10.1
    # via -r src/requirements.in
build==0.8.0
    # via pip-tools
certifi==2021.10.8
    # via
    #   requests
    #   sentry-sdk
charset-normalizer==2.0.12
    # via requests
click==8.1.2
    # via
    #   pip-tools
    #   rq
concurrent-log-handler==0.9.17
    # via -r src/requirements.in
croniter==1.3.4
    # via rq-scheduler
deprecated==1.2.13
    # via redis
django==3.2.13
    # via
    #   -r src/requirements.in
    #   django-cas-ng
    #   django-cors-headers
    #   django-debug-toolbar
    #   django-extensions
    #   django-filter
    #   django-redis
    #   django-rq
    #   djangorestframework
    #   drf-spectacular
django-cas-ng==4.1.1
    # via -r src/requirements.in
django-cors-headers==3.5.0
    # via -r src/requirements.in
django-debug-toolbar==3.1.1
    # via -r src/requirements.in
django-environ==0.4.5
    # via -r src/requirements.in
django-extensions==3.1.5
    # via -r src/requirements.in
django-filter==2.4.0
    # via -r src/requirements.in
django-redis==4.12.1
    # via -r src/requirements.in
django-rq==2.5.1
    # via -r src/requirements.in
djangorestframework==3.12.2
    # via
    #   -r src/requirements.in
    #   drf-spectacular
djangorestframework-api-key==2.2.0
    # via -r src/requirements.in
drf-spectacular==0.15.1
    # via -r src/requirements.in
gunicorn==20.0.4
    # via -r src/requirements.in
idna==3.3
    # via requests
importlib-resources==5.7.0
    # via jsonschema
inflection==0.5.1
    # via drf-spectacular
isodate==0.6.1
    # via rdflib
jsonschema==4.4.0
    # via drf-spectacular
lxml==4.8.0
    # via python-cas
packaging==21.3
    # via
    #   build
    #   redis
pep517==0.12.0
    # via build
pip-tools==6.7.0
    # via -r src/requirements.in
portalocker==2.4.0
    # via concurrent-log-handler
psutil==5.9.0
    # via rainbow-saddle
psycopg2-binary==2.8.6
    # via -r src/requirements.in
pyparsing==3.0.8
    # via
    #   packaging
    #   rdflib
pyrsistent==0.18.1
    # via jsonschema
python-cas==1.6.0
    # via django-cas-ng
python-dateutil==2.8.2
    # via
    #   croniter
    #   rq-scheduler
python-slugify==6.1.2
    # via -r src/requirements.in
pytz==2022.1
    # via
    #   babel
    #   django
pyyaml==6.0
    # via drf-spectacular
rainbow-saddle==0.4.0
    # via -r src/requirements.in
rdflib==5.0.0
    # via
    #   -r src/requirements.in
    #   skosmos-client
redis==4.2.2
    # via
    #   django-redis
    #   django-rq
    #   rq
requests==2.27.1
    # via
    #   python-cas
    #   skosmos-client
rq==1.10.1
    # via
    #   django-rq
    #   rq-scheduler
rq-scheduler==0.11.0
    # via -r src/requirements.in
sentry-sdk==1.31.0
    # via -r src/requirements.in
shortuuid==1.0.1
    # via -r src/requirements.in
six==1.16.0
    # via
    #   isodate
    #   python-cas
    #   python-dateutil
    #   rdflib
skosmos-client==0.3.0
    # via -r src/requirements.in
sqlparse==0.4.2
    # via
    #   django
    #   django-debug-toolbar
text-unidecode==1.3
    # via python-slugify
tomli==2.0.1
    # via
    #   build
    #   pep517
translitcodec==0.7.0
    # via -r src/requirements.in
uritemplate==4.1.1
    # via drf-spectacular
urllib3==1.26.17
    # via
    #   requests
    #   sentry-sdk
werkzeug==2.1.1
    # via -r src/requirements.in
wheel==0.37.1
    # via pip-tools
wrapt==1.14.0
    # via deprecated
zipp==3.8.0
    # via importlib-resources

# The following packages are considered to be unsafe in a requirements file:
# pip
# setuptools
